+++ spec/benchmarking_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
-require 'benchmark'
+require 'benchmark/ips'
 
 RSpec.describe ImageUtil::Benchmarking do
   it 'runs image creation benchmark' do
-    result = described_class.image_creation(1)
-    result.should be_a(Benchmark::Tms)
+    result = described_class.image_creation(0.1)
+    result.should be_a(Benchmark::IPS::Report)
   end
 end