+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Redimension filter to change image dimensions
 - Redimension attempts to resize the existing buffer when increasing height or
   adding a new dimension
-- Benchmarking subsystem with `bin/benchmark`
+- Benchmarking subsystem with `bin/benchmark` using `benchmark-ips`
 
 ## [0.2.0] - 2025-07-21
 - Ruby Sixel encoder now sets pixel aspect ratio metadata to display correctly in Windows Terminal